# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FAR Mail API
  version: 1.0.0
  description: |
    FAR Mail 的公开接口与自动化 API。服务器地址使用相对路径，始终指向
    当前部署来源，不包含写死的生产域名或 IP。

    `/api/v1` 仅接受专用 API Token 的 `Authorization: Bearer`；
    Admin Key 仅供 `/console/v1` 使用，不能替代 API Token。
servers:
  - url: /
    description: 当前部署来源
...
paths:
...
  /api/v1/mailboxes/cleanup:
    post:
      tags: [Mailboxes]
      operationId: cleanupMailboxes
      summary: 按条件清理邮箱
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
                domain: { type: string }
                only_expired: { type: boolean }
                only_empty: { type: boolean }
      responses:
        "200":
          description: 清理统计
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/emails/cleanup:
    post:
      tags: [Email]
      operationId: cleanupEmails
      summary: 按条件清理历史邮件
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                query: { type: string }
                domain: { type: string }
                older_than_minutes: { type: integer, minimum: 0 }
      responses:
        "200":
          description: 清理统计
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/lookup/mailbox:
    get:
      tags: [Lookup]
      operationId: lookupMailbox
      summary: 按完整地址查找邮箱
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: 邮箱信息
          content:
            application/json:
              schema:
                type: object
                properties:
                  mailbox:
                    $ref: "#/components/schemas/Mailbox"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/lookup/latest-code:
    get:
      tags: [Lookup]
      operationId: lookupLatestCode
      summary: 读取最新验证码
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: 验证码投影
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupCode"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lookup/latest-link:
    get:
      tags: [Lookup]
      operationId: lookupLatestLink
      summary: 读取最新验证链接
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: 验证链接投影
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LookupLink"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 32 lowercase hex characters
      description: 专用 API Token。不能使用 Admin Key，也不接受 X-API-Key 或 query-string key。
  parameters:
...
    Address:
      name: address
      in: query
      required: true
      description: 完整邮箱地址
      schema: { type: string, format: email }
  responses:
...
    Unauthorized:
      description: 缺少凭据或凭据无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Token scope 不允许该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或当前 Token 不可访问
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        success: { type: boolean, const: false }
        error: { type: string }
        message: { type: string }
        error_code: { type: string }
        status: { type: integer }
        request_id: { type: string, format: uuid }
...
    Mailbox:
      type: object
      required: [id, address, domain_id, full_address, keep_forever, created_at]
      properties:
        id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
        address: { type: string }
        domain_id: { type: integer, format: int64 }
        full_address: { type: string, format: email }
        expires_at: { type: [string, "null"], format: date-time }
        keep_forever: { type: boolean }
        email_count: { type: integer }
        latest_received_at: { type: [string, "null"], format: date-time }
        latest_code: { type: string }
        latest_link: { type: string }
        created_at: { type: string, format: date-time }
...
    LookupCode:
      type: object
      properties:
        mailbox:
          $ref: "#/components/schemas/Mailbox"
        email_id: { type: string, format: uuid }
        sender: { type: string }
        subject: { type: string }
        received_at: { type: string, format: date-time }
        code: { type: string }
        matched_by: { type: string }
        has_code: { type: boolean }
    LookupLink:
      type: object
      properties:
        mailbox:
          $ref: "#/components/schemas/Mailbox"
        email_id: { type: string, format: uuid }
        sender: { type: string }
        subject: { type: string }
        received_at: { type: string, format: date-time }
        link: { type: string, format: uri }
        matched_by: { type: string }
        has_link: { type: boolean }
    DeleteResult:
      type: object
      additionalProperties: true
      properties:
        message: { type: string }
        deleted: { type: integer }